feat(cva): add config and utils entry points
