+++ CONTRIBUTING.md
@@ -25,7 +25,7 @@ focused:
 - **Commits** follow the ["Conventional Commits" specification](https://www.conventionalcommits.org/en/v1.0.0/). This allows for changelogs to be generated automatically upon release.
 - **Code** is formatted via [Prettier](https://prettier.io/)
 - **JavaScript** is written as [TypeScript](https://www.typescriptlang.org/) where possible.
-- **`packages/cva`'s type exports**: any type that can appear _named_ (not structurally expanded) in a consumer's generated `.d.ts` when they compile with `declaration: true` must be `export`ed from `src/index.ts`, even if it's not meant for direct use — an unexported-but-nameable type breaks their build with a `TS4023`/`TS2459`-family error even though `cva`'s own build stays green. Not every type reachable from a public signature needs this: TypeScript structurally expands some of them (e.g. the call-signature parameter helpers) instead of naming them, so those stay unexported on purpose — export the minimum that a real consumer build fails to compile without (see `AGENTS.md`'s Learnings for how to check). These exports exist for that portability reason only, not as a feature we want people to reach for directly, so mark them with a short JSDoc saying so and don't add docs-site coverage for them. `packages/cva/src/index.test.ts` pins the current set, but only catches _losing_ one of these exports, not a new type that newly needs one.
+- **`packages/cva`'s type exports**: any type that can appear _named_ (not structurally expanded) in a consumer's generated `.d.ts` when they compile with `declaration: true` must be `export`ed from the public entry point that exposes the signature (`cva`, `cva/config`, or `cva/utils`), even if it's not meant for direct use — an unexported-but-nameable type breaks their build with a `TS4023`/`TS2459`-family error even though `cva`'s own build stays green. Not every type reachable from a public signature needs this: TypeScript structurally expands some of them (e.g. the call-signature parameter helpers) instead of naming them, so those stay unexported on purpose — export the minimum that a real consumer build fails to compile without (see `AGENTS.md`'s Learnings for how to check). These exports exist for that portability reason only, not as a feature we want people to reach for directly, so mark them with a short JSDoc saying so and don't add docs-site coverage for them. `packages/cva/src/index.test.ts` pins the current set, but only catches _losing_ one of these exports, not a new type that newly needs one.
 
 ## Getting Started
 
@@ -80,7 +80,7 @@ Both published packages (`packages/cva` and `packages/class-variance-authority`)
 
 #### How the packages transform for publish
 
-The `exports` and `publishConfig.exports` blocks in each package's `package.json` are **machine-generated**: the config's `exports: { devExports: true }` makes tsdown rewrite both on every build. **Never hand-edit them** — the next build silently overwrites your change; adjust that package's `tsdown.config.mts` (or the shared base) instead. (One exception: `class-variance-authority`'s `publishConfig.typesVersions` — the node10 fallback for its `./types` subpath — is hand-maintained; tsdown doesn't generate it but preserves it across rebuilds.) The two blocks implement a dev/publish split:
+The `exports` and `publishConfig.exports` blocks in each package's `package.json` are **machine-generated**: the config's `exports: { devExports: true }` makes tsdown rewrite both on every build. **Never hand-edit them** — the next build silently overwrites your change; adjust that package's `tsdown.config.mts` (or the shared base) instead. The hand-maintained exception is `publishConfig.typesVersions`: `class-variance-authority` has a node10 fallback for `./types`, and `cva` has fallbacks for `./config` and `./utils`. tsdown preserves them but does not generate them. The two blocks implement a dev/publish split:
 
 - The top-level `exports` points at `./src/*.ts`, so workspace consumers (tests, examples, docs) always resolve the raw TypeScript source with no build step in between.
 - `publishConfig.exports` points at `dist/`. pnpm applies `publishConfig` when packing or publishing (`pnpm pack` / `pnpm publish` — never `npm pack`, which skips the rewrite entirely), so the tarball people install resolves the built output.
