+++ REVIEW.md
@@ -10,7 +10,7 @@ Two facts set the stakes: both packages are published to npm and depended on by 
 
 The highest-stakes surface in the repo. Types are the product here as much as the runtime is.
 
-- **A newly-nameable internal type must be exported.** Per [`CONTRIBUTING.md`](./CONTRIBUTING.md#style-guide): a type that appears _named_ in a consumer's generated `.d.ts` under `declaration: true` breaks their build with a `TS4023`/`TS2459`-family error if it isn't exported from `src/index.ts` — while `cva`'s own build stays green. The `describe("exported types", …)` and `describe("CVAVariantShape", …)` blocks in `packages/cva/src/index.test.ts` pin the _existing_ portability exports (`CVAComponent`, `CVAComponentShape`, `CVAVariantShape`), so losing one fails `pnpm --filter cva check:tsc`. Nothing detects a _newly_ required one. If a PR reshapes a public signature, ask for the packed-artifact verification documented in [`AGENTS.md`](./AGENTS.md#learnings) — `pnpm pack` (never `npm pack`), extract, compile a real consumer with `--module nodenext`.
+- **A newly-nameable internal type must be exported.** Per [`CONTRIBUTING.md`](./CONTRIBUTING.md#style-guide): a type that appears _named_ in a consumer's generated `.d.ts` under `declaration: true` breaks their build with a `TS4023`/`TS2459`-family error if it isn't exported from the public entry point that exposes the signature (`cva`, `cva/config`, or `cva/utils`) — while `cva`'s own build stays green. The `describe("exported types", …)` and `describe("CVAVariantShape", …)` blocks in `packages/cva/src/index.test.ts` pin the _existing_ portability exports (`CVAComponent`, `CVAComponentShape`, `CVAVariantShape`), so losing one fails `pnpm --filter cva check:tsc`. Nothing detects a _newly_ required one. If a PR reshapes a public signature, ask for the packed-artifact verification documented in [`AGENTS.md`](./AGENTS.md#learnings) — `pnpm pack` (never `npm pack`), extract, compile a real consumer with `--module nodenext`.
 - Not everything reachable needs exporting — TypeScript structurally expands some helpers instead of naming them. Export the minimum a real consumer build fails without, mark it with the same "you shouldn't really use it directly" JSDoc the existing three carry, and don't add docs-site coverage for it.
 - **Inference can widen silently and still compile.** `pnpm check` only proves `tsc` is happy. Whether `VariantProps<typeof x>` still resolves to a literal union rather than `string`, or `getSchema` still names every variant, is pinned only by the `expectTypeOf(…).toEqualTypeOf<…>()` assertions in the test files. Treat a loosened assertion (`toEqualTypeOf` → `toMatchTypeOf`), a deleted one, or a deleted `@ts-expect-error` as part of the diff to review, not as test housekeeping.
 - **The awkward generics are load-bearing, and each says why in a comment.** `ComposedTuple` splits inference across two type parameters so an array literal stays a tuple instead of collapsing to a union (which silently drops structurally-subtyped components). `RightMerge` is a mapped type rather than `Omit<A, keyof B> & B` because the latter stays deferred and breaks `any`-narrowing downstream. `DefaultsOf` normalises `undefined` to `{}` because `keyof never` is `string | number | symbol` and poisons the key union. `CVAComponentShape` takes `any` arguments on purpose — a shaped `config` rejects every real component via props contravariance. A PR that "simplifies" any of these should demonstrate the documented failure no longer applies.
@@ -36,12 +36,12 @@ The highest-stakes surface in the repo. Types are the product here as much as th
 CI's `build` job runs tsdown's `publint`/`attw`/`unused` gates, so a genuinely _broken_ publish shape fails on its own. What it can't see:
 
 - **A hand-edit to `exports` or `publishConfig.exports` that happens to be valid.** Both blocks are regenerated on every build (`exports: { devExports: true }` in [`.config/tsdown.base.mts`](./.config/tsdown.base.mts)), so the edit is silently reverted by the next build and the intent is lost. A diff to either block with no matching change to that package's `tsdown.config.mts` or the shared base is the tell.
-- **Published `typesVersions` fallbacks** — `class-variance-authority`'s `./types` and `cva`'s `./core` node10 fallbacks — are hand-maintained. tsdown preserves them but doesn't generate them, so a deletion never comes back.
+- **Published `typesVersions` fallbacks** — `class-variance-authority`'s `./types` and `cva`'s `./config` and `./utils` node10 fallbacks — are hand-maintained. tsdown preserves them but doesn't generate them, so a deletion never comes back.
 - **The published packages deliberately omit `engines.node`** so they don't constrain consumers ([`CONTRIBUTING.md`](./CONTRIBUTING.md#nodejs-versions)). syncpack treats the field as optional and only snaps a declared one to the root, so an added `"24"` sails through — still wrong for a published package.
 
 ## Size and dependency budget
 
-- `bundlesize` enforces the numbers, not the choice of numbers. The `size-limit` blocks cap the `cva` preset at `1.31KB`, `cva/core` at `1.05KB`, and `class-variance-authority` at `1.2KB`. A PR that raises its own limit to fit gets a green job — that is a review decision, weighed against the "performance & minimal footprint" project goal.
+- `bundlesize` enforces the numbers, not the choice of numbers. The `size-limit` blocks cap the `cva` preset at `1.35KB`, `cva/config` at `1.05KB`, `cva/utils` at `0.35KB`, and `class-variance-authority` at `1.2KB`. A PR that raises its own limit to fit gets a green job — that is a review decision, weighed against the "performance & minimal footprint" project goal.
 - `clsx` is the only runtime dependency of either package. Anything added to `dependencies` under `packages/*` is a headline change, not a detail.
 - [`pnpm-workspace.yaml`](./pnpm-workspace.yaml) carries two supply-chain protections that a PR can quietly weaken: `minimumReleaseAge`, and the explicit `allowBuilds` allowlist (adding a package there grants it install-time script execution). Removing or lowering `minimumReleaseAge` to work around an `ERR_PNPM_MISSING_TIME` is the specific anti-fix called out in [`AGENTS.md`](./AGENTS.md#learnings) — the right answer is to re-run the install.
 - Dependabot is configured for `github-actions` only ([`.github/dependabot.yml`](./.github/dependabot.yml)), so npm bumps arrive by hand and carry no automated provenance.
