+++ examples/beta/react-with-cn/src/cva.config.ts
@@ -1,4 +1,4 @@
-import { defineConfig } from "cva/core";
+import { defineConfig } from "cva/config";
 import { cn } from "cn";
 
 export const { cva, cx } = defineConfig({ cx: cn });
