+++ packages/cva/tsdown.config.mts
@@ -3,6 +3,6 @@ import { base } from "../../.config/tsdown.base.mts";
 
 export default defineConfig({
   ...base,
-  // Hand-maintained node10 `typesVersions` fallback for `./core`.
-  entry: ["src/index.ts", "src/core.ts"],
+  // Hand-maintained node10 `typesVersions` fallbacks cover these subpaths.
+  entry: ["src/index.ts", "src/config.ts", "src/utils.ts"],
 });
